MTL-1561 Add GCP mode for supporting GCP NVME (#15)
This new toggle will insert the partitioning prefix into all disks,
assuming that the disks are GCP provided NVME.

This implicitly means that only NVME is supported when gcp-mode is
engaged. We can add logic to handle our full-suite of busses, only if
we have a use in SAS and SATA GCP disks.