MTL-1811 Bad XFS Option Prevents Mounting in SLES15SP4
This option must be to a power-of-2, for some time the allocsize
optimization has not been to a power-of-2. Somehow this incorrect option
has gone unoticed and XFS mounts have been working fine in both
SLES15SP2 and SLES15SP3. However in SLES15SP4 our luck has ran out, this
incorrect value (incorrect because it isn't a power of 2) prevents XFS
from mounting.

It would seem that this value was incorrectly added a long time ago but
had gone unnoticed, most XFS optimization guides suggest 131072(k) and
our value is missing that trailing "2". 131072 is 2 to the power of 17
(in other words, 131072 log2 = 17). 13107 log2 is not an integer.