CASMINST-5103 Better LVM logging (#48)
* CASMINST-5103 Better LVM logging

The pave function doesn't report to the console whether or not
any volume groups were found, let alone which logical volumes were
removed.

This change adds discovered volume groups and removed items to the
console.

* CASMINST-5103 Fatal if volume group still exists

If any given volume group that was deleted still exists, fail the boot
with an error. In the error message mention next steps.