Remove default LIVE scheme and authority
Along the same theme as the previous commit, remove the chance for code to run against non-existent drive labels by removing the default/assumed labels.

Tidies up the cases for SQFSRAID; adds a blank case for safely loading `metal-md-lib.sh` where `root` is something else, and prints messages similar to the `oval_drive_scheme` case.

This also fixes a hidden bug on line 87 of `metal-md-lib.sh`, this line should be only overriding `SQFS_DRIVE_SCHEME` if it is equal to `CDLABEL` but the `||` was doing the opposite of that.