Run scheme and authority parsing once
`metal-kdump.sh` was recreating some of the logic in `metal-md-lib.sh` for resolving the overlayFS scheme and authority. Short of sourcing `metal-md-lib.sh`, it would bode well if we parsed the scheme and authority every boot and only once.

By moving the scheme and authority parsing out of `metal-md-lib.sh`, the parsing will occur once and at the beginning of the boot instead of every time `metal-md-lib.sh` is sourced.

Additionally, this removes the `exit` calls from `metal-md-lib.sh`. Libraries shouldn't exit.