capitalization and extraneous check
`MOUNT_OVERLAYfs_id` should be all lowercase, it was recently awkward cased in a previous, greedy string replace operation.

There is also no point in checking if `metal_overlayfs_id` is empty, it will always be empty until we set it. We want this variable to always be fresh, it should always recalculate when the function is ran later.